[1-1] BLOG - GET Requests
[Net Ninja] VueJS2 Tutorial #33 :
BLOG - GET Requests

https://youtu.be/aoWqFLGCK60
